Add controller for full application
DRY up a few bits in the process
