+++ config/routes.rb
@@ -9,5 +9,6 @@ Rails.application.routes.draw do
   
   resources :applications do
     resources :eligibility, controller: 'applications/eligibility'
+    resources :full_application, controller: 'applications/full_application'
   end
 end
